requirements.txt
google-generativeai
fastapi
uvicorn
websockets
redis
supabase
twilio
httpx
pydantic
pydantic-settings
python-dotenv
loguru
tenacity
beautifulsoup4
lxml
phonenumbers
psycopg2-binary
google-cloud-bigquery
google-cloud-firestore
requests
SpeechRecognition
pydub
google-adk
google-cloud
fastapi[standard]